requirements.txt

# Core Django Framework
django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=24.1

# Database & ORM
psycopg2-binary>=2.9.9
sqlparse>=0.4.4

# Environment & Configuration
python-dotenv>=1.0.0

# Production Server & Deployment
gunicorn>=21.2.0
whitenoise>=6.6.0

# Machine Learning & AI
tensorflow>=2.13.0
tensorflow-hub>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# API & Data Processing
requests>=2.31.0
google-generativeai>=0.3.0
Pillow>=10.0.0

# Utilities
celery>=5.3.0
redis>=5.0.0
pytest>=7.4.0
pytest-django>=4.7.0
